+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### Centilator 2019 [Alpha 1.0] - 9/2/2018
+* [ADDED] New Homepage
+* [CHANGED] Now the total/answer is inside the windows
+* [ADDED] Calulator tape to the Basic Calculator
+* [ADDED] Basic function to the tape
+* [CHANGED] Aligned the Calculator button exactly perfect.
+* [ADDED] Colors to the buttons
+
+
 ### Centilator 2019 [Indev 2.0] - 8/31/2018
 * Has basic feutures set!
 * Will move to alpha stage